create asyncThunk deleteFromFavorite

+++ src/Components/UserProfile/NavComponents/ActionsComponent/ActionsComponent.js
@@ -1,40 +1,23 @@
 import React from 'react';
 import ActionButton from '../../../Action Bar/ActionButton';
-import { AiFillHeart, AiFillStar, AiOutlineDelete, AiOutlineUnorderedList } from 'react-icons/ai';
+import {AiFillHeart, AiFillStar, AiOutlineDelete, AiOutlineUnorderedList} from 'react-icons/ai';
 import styles from '../../UserProfile.module.scss';
-import { db } from '../../../../firebase';
-import { useDispatch, useSelector } from 'react-redux';
-import { collection, deleteDoc, getDocs, query, where } from 'firebase/firestore';
-import { removeFavorite } from '../../../../redux/slices/favoriteSlice';
+import {useDispatch, useSelector} from 'react-redux';
+import {deleteFavorite, removeFavorite} from '../../../../redux/slices/favoriteSlice';
 
-const ActionsComponent = ({ movieId , removeFromFavorites }) => {
+const ActionsComponent = ({movieId}) => {
   const userId = useSelector((state) => state.user.id);
   const dispatch = useDispatch();
+  const isFavorite = useSelector((state) => state.favorites.isFavorite[movieId])
 
-  const removeCollectionByMovieId = async (movieId) => {
-
-    const collectionRef = collection(db, `users/${userId}/favorites`);
-    const queryRef = query(collectionRef, where('movieId', '==', movieId));
-    const snapshot = await getDocs(queryRef);
-
-    snapshot.forEach((doc) => {
-      deleteDoc(doc.ref);
-    });
-
-    dispatch(removeFavorite(movieId));
-    console.log('Collection removed for movieId:', movieId);
-
-    removeFromFavorites(movieId);
-  };
-
-  const removeFromFavoriteHandle = async () => {
-    try {
-      await removeCollectionByMovieId(movieId);
-    } catch (error) {
-      console.error('Error removing collection for movieId:', error);
+  const removeFromFavoriteHandle = () => {
+    if (userId) {
+      dispatch(deleteFavorite({ userId, movieId }));
+      dispatch(removeFavorite(movieId));
+    } else {
+      alert('User data has not loaded yet');
     }
   };
-
 
   return (
 
@@ -53,8 +36,10 @@ const ActionsComponent = ({ movieId , removeFromFavorites }) => {
       </div>
       <div>
         <ActionButton
+          onClick={removeFromFavoriteHandle}
           icon={<AiFillHeart
             size={30}
+            color={isFavorite ? 'red' : null}
           />}/>
         <p>Favorite</p>
       </div>
